requirements.txt
fastapi
uvicorn
pdfminer.six
sentence-transformers
faiss-cpu
python-multipart
